Don't nop out playerSpeed instruction
This instruction is important to keep, because Marisa's
bombs use the playerSpeed variable. (deleting this instruction
causes Marisa to move slow forever after bombing!)

Thankfully, it appears that ECL scripts run after this instruction,
but before the player moves, so our playerSetSpeedMult can still
work! (it just needs to be called every frame)